# perf(stdlib): optimize limit transformation