The limit transformation has been optimized to use a passthrough dataset
and to use arrow slices rather than copying the input to a new table.
This reduces the number of memory allocations which makes the
transformation use a more consistent amount of memory and speeds up the
transformation because it doesn't have to wait for memory allocations to
finish.

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written