@jsternberg 2020-01-03
Benchmarks:

```
benchmark                       old ns/op      new ns/op     delta
BenchmarkLimit_1N_1000-12       608936123      240179582     -60.56%
BenchmarkLimit_100N_1000-12     1527333381     237654834     -84.44%

benchmark                       old allocs     new allocs     delta
BenchmarkLimit_1N_1000-12       3671443        810008         -77.94%
BenchmarkLimit_100N_1000-12     4691770        810008         -82.74%

benchmark                       old bytes      new bytes     delta
BenchmarkLimit_1N_1000-12       282523411      58560502      -79.27%
BenchmarkLimit_100N_1000-12     1027038739     58560394      -94.30%
```